refactor manager sandbox service split
